doc: added details around console.timeEnd changes
Since PR https://github.com/nodejs/node/pull/29251 is merged into
master, the implementation of `console.timeEnd` is not to print the
milliseconds on the stdout but instead print a more readable time unit
on the console. For example if the total time comes out to be 3456
milliseconds, it will be printed as 3.45 seconds so that it is much more
readable and consistent.

PR-URL: https://github.com/nodejs/node/pull/35027
Fixes: https://github.com/nodejs/node/issues/35011
Refs: https://github.com/nodejs/node/pull/29251
Reviewed-By: Benjamin Gruenbaum <benjamingr@gmail.com>
Reviewed-By: Anna Henningsen <anna@addaleax.net>
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>