src: rename AliasedBufferInfo->AliasedBufferIndex
This commit suggest renaming AlaisedBufferInfo to AlaisedBufferIndex to
make the code more readable.

The main motivation for this change is that I personally think that
the following code could be a little clearer:

  context->GetDataFromSnapshotOnce<V8T>(*info_).ToLocalChecked();

Even knowing that GetDataFromSnapshotOnce takes a size_t I had to double
check the type of info_ to make sure.

PR-URL: https://github.com/nodejs/node/pull/36339
Reviewed-By: Zeyu Yang <himself65@outlook.com>
Reviewed-By: Rich Trott <rtrott@gmail.com>