doc: revise objections section of collaborator guide
The bold text is overused in this section. This change removes it. It
also changes some wording: PR -> pull request, can -> may in a rare
instance where "may" is the right choice in our docs, etc. (The
recommendation to avoid "may" is because "may" implies permission rather
than just ability. Yes, I'm probably the one that changed it from "may"
in the first place.)

PR-URL: https://github.com/nodejs/node/pull/37525
Reviewed-By: Antoine du Hamel <duhamelantoine1995@gmail.com>
Reviewed-By: Michaël Zasso <targos@protonmail.com>
Reviewed-By: Benjamin Gruenbaum <benjamingr@gmail.com>
Reviewed-By: Darshan Sen <raisinten@gmail.com>
Reviewed-By: Gerhard Stöbich <deb2001-github@yahoo.de>